requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
diffusers==0.32.2
efficientnet_pytorch==0.7.1
einops==0.8.0
eval_type_backport==0.2.0
evaluation==0.0.2
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
glog==0.3.1
huggingface-hub==0.27.0
idna==3.10
image-quality==1.2.7
imageio==2.35.1
imgaug==0.2.6
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
libsvm==3.23.0.4
lightning==2.3.3
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
networkx==3.1
nose==1.3.7
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
pillow==10.4.0
pretrainedmodels==0.7.4
progress==1.6
propcache==0.2.0
protobuf==5.29.1
pybrisque==1.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.1.4
python-gflags==3.1.2
pytorch-lightning==2.4.0
pytorch_ssim==0.1
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
sewar==0.4.6
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.2
tokenizers==0.20.3
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
urllib3==2.2.3
yarl==1.15.2
zipp==3.20.2
